requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==4.2.27
django-cors-headers==4.9.0
django-sslserver==0.22
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==11.3.0
psycopg==3.2.13
PyJWT==2.10.1
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.4
typing_extensions==4.15.0
Unidecode==1.4.0
uritemplate==4.2.0
